Extract out parsing methods to slider controller from math utils
